+++ apps/dashboard/src/lib/components/Org/LandingPage/Templates/Experts/Home.svelte
@@ -1,15 +1,15 @@
 <script>
   import { t } from '$lib/utils/functions/translations';
-  import Courses from './Courses.svelte';
-  import Faq from './Faq.svelte';
-  import Footer from './Footer.svelte';
-  import FooterNote from './FooterNote.svelte';
-  import Hero from './Hero.svelte';
-  import Learn from './Learn.svelte';
-  import LearningPath from './LearningPath.svelte';
-  import Navigation from './Navigation.svelte';
+  import Courses from '$lib/components/Org/LandingPage/Templates/Experts/components/Courses.svelte';
+  import Faq from '$lib/components/Org/LandingPage/Templates/Experts/components/Faq.svelte';
+  import Footer from '$lib/components/Org/LandingPage/Templates/Experts/components/Footer.svelte';
+  import FooterNote from '$lib/components/Org/LandingPage/Templates/Experts/components/FooterNote.svelte';
+  import Hero from '$lib/components/Org/LandingPage/Templates/Experts/components/Hero.svelte';
+  import Learn from '$lib/components/Org/LandingPage/Templates/Experts/components/Learn.svelte';
+  import LearningPath from '$lib/components/Org/LandingPage/Templates/Experts/components/LearningPath.svelte';
+  import Navigation from '$lib/components/Org/LandingPage/Templates/Experts/components/Navigation.svelte';
   import PageLoader from '$lib/components/Org/LandingPage/PageLoader.svelte';
-  import Testimonial from './Testimonial.svelte';
+  import Testimonial from '$lib/components/Org/LandingPage/Templates/Experts/components/Testimonial.svelte';
 
   export let org = {};
 </script>
