+++ apps/dashboard/src/lib/components/Org/LandingPage/Templates/Experts/components/Courses.svelte
@@ -1,7 +1,7 @@
 <script>
-  import CourseCard from '../../components/CourseCard.svelte';
+  import CourseCard from '$lib/components/Org/LandingPage/components/CourseCard.svelte';
   import PrimaryButton from '$lib/components/PrimaryButton/index.svelte';
-  import { landingPageSettings } from '../../../Settings/store';
+  import { landingPageSettings } from '$lib/components/Org/Settings/store';
   import CoursesEmptyIcon from '$lib/components/Icons/CoursesEmptyIcon.svelte';
   import Box from '$lib/components/Box/index.svelte';
   import { t } from '$lib/utils/functions/translations';
@@ -22,7 +22,7 @@
         <CardLoader />
       </div>
     {:else if $courses.length > 0}
-      <section class="flex flex-wrap gap-4 p-4">
+      <section class="flex flex-wrap items-center justify-center md:justify-start gap-4 p-4">
         {#each $courses.slice(0, viewAll ? $courses.length : 3) as courseData}
           <CourseCard
             className="bg-[#FDFDFD]"
@@ -37,7 +37,11 @@
       </section>
       {#if $courses.length > 3}
         <div class="w-full flex items-center justify-center my-5">
-          <PrimaryButton label="VIEW MORE COURSES" className="rounded-none text-lg" />
+          <PrimaryButton
+            label="VIEW MORE COURSES"
+            className="rounded-none text-lg"
+            onClick={() => (viewAll = !viewAll)}
+          />
         </div>
       {/if}
     {:else}
