+++ apps/dashboard/src/lib/components/Org/LandingPage/Templates/Experts/components/Testimonial.svelte
@@ -4,7 +4,7 @@
 
 <section class="px-4 pt-4 pb-20 h-full bg-white">
   <h1 class="text-center text-3xl">Student testimonial</h1>
-  <section class="flex flex-wrap gap-4 p-4">
+  <section class="flex flex-wrap items-center justify-center md:justify-start gap-4 p-4">
     <TestimonialCard />
     <TestimonialCard />
     <TestimonialCard />
