+++ apps/dashboard/src/lib/components/Org/LandingPage/Templates/Plain/Course.svelte
@@ -1,0 +1,1 @@
+<p>plan course</p>
